+++ CHANGES.md
@@ -12,6 +12,7 @@
 
 ##### Additions :tada:
 
+- Improved property change checks in `Cesium3DTileset::LoadTileset`. 
 - Added macOS support.
 - Added support for the legacy `gltfUpAxis` property in a tileset `asset` dictionary. Although this property is **not** part of the specification, there are many existing assets that use this property and had been shown with a wrong rotation otherwise.
 - Changed the log level for the tile selection output from `Display` to `Verbose`. With default settings, the output will no longer be displayed in the console, but only written to the log file.